requirements.txt
django >= 2.1.6, < 3.0


django-simple-rest

#Django==3.2.20

djongo==1.3.6
sqlparse==0.2.4
